Fraction分数计算优化:线程安全、溢出处理、功能完善、测试完善

+++ src/test/java/com/tingfeng/util/java/base/common/bean/fraction/IntFractionTest.java
@@ -155,7 +155,7 @@ public class IntFractionTest {
         int multipleValue = RandomUtils.randomInt(100);
         IntFraction fraction = new IntFraction( multipleValue , baseValue * multipleValue).simpleFraction();
         System.out.println(fraction.getValue());
-        Assert.assertEquals(Math.abs(baseValue), fraction.getDenominator());
+        Assert.assertEquals(Math.abs(baseValue), fraction.getDenominator().intValue());
     }
 
     @Test
